pyproject.toml
[project]
name = "stable-emotion-dubbing-indextts2"
version = "0.1.0"
description = "Subtitle-driven multi-speaker emotion dubbing pipeline for IndexTTS2."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydub>=0.25.1",
  "PyYAML>=6.0.1",
  "jiwer>=3.0.4",
]

[project.optional-dependencies]
eval = [
  "faster-whisper>=1.0.0",
  "openai-whisper>=20231117",
  "speechbrain>=1.0.0",
  "torch",
  "torchaudio",
]
test = [
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["stable_dubbing*"]
